23.03.2021

+++ Core/Utilities/Helpers/FileHelper.cs
@@ -7,86 +7,39 @@ using System.Text;
 
 namespace Core.Utilities.Helpers
 {
-    public class FileHelper
+    public class FileHelper:IFileHelper
     {
-        public static string newPath(IFormFile file)
+        public string Add(IFormFile file, string path)
         {
-            System.IO.FileInfo ff = new System.IO.FileInfo(file.FileName);
-            string fileExtension = ff.Extension;
+            var sourcepath = Path.GetTempFileName();
 
-            var creatingUniqueFilename = Guid.NewGuid().ToString("N")
-                + "_" + DateTime.Now.Month
-                + "_" + DateTime.Now.Day
-                + "_" + DateTime.Now.Year + fileExtension;
-
-            string path = Path.Combine(Directory.GetParent(System.IO.Directory.GetCurrentDirectory()).FullName + @"\Images");
-
-            string result = $@"{path}\{creatingUniqueFilename}";
-
-            return result;
-        }
-
-        public static string AddAsync(IFormFile file)
-        {
-            var result = newPath(file);
-
-            try
-            {
-                var sourcePath = Path.GetTempFileName();
-                if (file.Length > 0)
+            if (file.Length > 0)
+                using (var stream = new FileStream(sourcepath, FileMode.Create))
                 {
-                    using (var stream = new FileStream(sourcePath, FileMode.Create))
-                    {
-                        file.CopyTo(stream);
-                    }
+                    file.CopyTo(stream);
                 }
 
-                File.Move(sourcePath, result);
-            }
-            catch (Exception exception)
-            {
-                return exception.Message;
-            }
-
-            return result;
+            File.Move(sourcepath, path);
+            return path;
         }
 
-        public static string UpdateAsync(string sourcePath, IFormFile file)
-        {
-            var result = newPath(file);
+        public string Update(string pathToUpdate, IFormFile file, string path)
 
-            try
-            {
-                if (sourcePath.Length > 0)
+        {
+            if (path.Length > 0)
+                using (var stream = new FileStream(path, FileMode.Create))
                 {
-                    using (var stream = new FileStream(result, FileMode.Create))
-                    {
-                        file.CopyTo(stream);
-                    }
+                    file.CopyTo(stream);
                 }
 
-                File.Delete(sourcePath);
-            }
-            catch (Exception exception)
-            {
-                return exception.Message;
-            }
-
-            return result;
+            File.Delete(pathToUpdate);
+            return path;
         }
 
-        public static IResult DeleteAsync(string path)
+        public void Delete(string path)
         {
-            try
-            {
+            if (File.Exists(path))
                 File.Delete(path);
-            }
-            catch (Exception exception)
-            {
-                return new ErrorResult(exception.Message);
-            }
-
-            return new SuccessResult();
         }
     }
 }
